+++ FoodSpyAPI/Interfaces/IMealFood.cs
@@ -3,7 +3,7 @@ namespace FoodSpyAPI.Interfaces
 {
 	public interface IMealFood
 	{
-		float Quantity { get; set; }
+		double Quantity { get; set; }
 
 		string Unit { get; set; }
 	}
